+++ domain/enums/activity_type.py
@@ -1,13 +1,4 @@
-import enum
+from data.model.activity_type import ActivityTypeDbo
 
 
-class ActivityType(enum.Enum):
-    """
-    Enum for Activity Type
-    """
-    EVENT = 'EVENT'
-    MAINTENANCE = 'MAINTENANCE'
-
-    PRACTICE = 'PRACTICE'
-    LABOUR = 'LABOUR'
-    IDLE = 'IDLE'
+ActivityType = ActivityTypeDbo